Move ui-related client methods to ui clients
