+++ credentialdigger/client.py
@@ -18,9 +18,6 @@ Repo = namedtuple('Repo', 'url last_scan')
 Discovery = namedtuple(
     'Discovery',
     'id file_name commit_id line_number snippet repo_url rule_id state timestamp')
-FilesSummary = namedtuple(
-    'FilesSummary',
-    'file_name tot_discoveries new false_positives addressing not_relevant')
 
 
 class Interface(ABC):
@@ -359,33 +356,6 @@ class Client(Interface):
             all_discoveries.append(dict(Discovery(*result)._asdict()))
             result = cursor.fetchone()
         return all_discoveries
-
-    def get_files_summary(self, query, repo_url):
-        """ Get aggregated discoveries info on all files of a repository.
-
-        Parameters
-        ----------
-        repo_url: str
-            The url of the repository
-
-        Returns
-        -------
-        list
-            A list of files with aggregated data (dictionaries)
-
-        Raises
-        ------
-            TypeError
-                If any of the required arguments is missing
-        """
-        cursor = self.db.cursor()
-        files = []
-        cursor.execute(query, (repo_url,))
-        result = cursor.fetchone()
-        while result:
-            files.append(dict(FilesSummary(*result)._asdict()))
-            result = cursor.fetchone()
-        return files
 
     def get_discovery(self, query, discovery_id):
         """ Get a discovery.
