+++ credentialdigger/client_postgres.py
@@ -264,32 +264,6 @@ class PgClient(Client):
             file_name=file_name,
             query=query)
 
-    def get_files_summary(self, repo_url):
-        """ Get aggregated discoveries info on all files of a repository.
-
-        Parameters
-        ----------
-        repo_url: str
-            The url of the repository
-
-        Returns
-        -------
-        list
-            A list of files with aggregated data (dictionaries)
-        """
-        return super().get_files_summary(
-            repo_url=repo_url,
-            query=(
-                "SELECT file_name,"
-                " COUNT(*) AS tot_discoveries,"
-                " COUNT(CASE WHEN state='new' THEN 1 END) AS new,"
-                " COUNT(CASE WHEN state='false_positive' THEN 1 END) AS false_positives,"
-                " COUNT(CASE WHEN state='addressing' THEN 1 END) AS addressing,"
-                " COUNT(CASE WHEN state='not_relevant' THEN 1 END) AS not_relevant"
-                " FROM discoveries WHERE repo_url=%s"
-                " GROUP BY file_name"
-            ))
-
     def get_discovery(self, discovery_id):
         """ Get a discovery.
